# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/change-password:
    post:
      operationId: changeAdminPassword
      tags: [admin]
      summary: Change admin password (used for forced first-login change too)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChangeInput"
      responses:
        "204":
          description: Changed
        "400":
          description: Validation error
        "401":
          description: Not logged in / wrong current password
  /admin/dashboard-summary:
    get:
      operationId: getAdminDashboardSummary
      tags: [admin]
      summary: Get admin dashboard summary metrics
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDashboardSummary"
        "401":
          description: Not logged in
  /admin/students:
    get:
      operationId: listStudents
      tags: [admin]
      summary: List students (search/filter/sort)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: className
          in: query
          schema:
            $ref: "#/components/schemas/ClassLevel"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/StudentStatus"
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, oldest, nameAsc, nameDesc]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
        "401":
          description: Not logged in
    post:
      operationId: createStudent
      tags: [admin]
      summary: Manually add a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentCreateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          description: Validation error
        "401":
          description: Not logged in
  /admin/students/{id}:
    get:
      operationId: getStudent
      tags: [admin]
      summary: Get a student by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "401":
          description: Not logged in
        "404":
          description: Not found
    patch:
      operationId: updateStudent
      tags: [admin]
      summary: Update a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "401":
          description: Not logged in
        "404":
          description: Not found
    delete:
      operationId: deleteStudent
      tags: [admin]
      summary: Permanently delete a terminated student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not logged in
        "404":
          description: Not found
  /admin/students/{id}/terminate:
    post:
      operationId: terminateStudent
      tags: [admin]
      summary: Terminate a student (moves to terminated list, 60-day retention)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TerminationInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "401":
          description: Not logged in
        "404":
          description: Not found
  /admin/students/{id}/restore:
    post:
      operationId: restoreStudent
      tags: [admin]
      summary: Restore a terminated student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "401":
          description: Not logged in
        "404":
          description: Not found
  /admin/terminated-students:
    get:
      operationId: listTerminatedStudents
      tags: [admin]
      summary: List terminated students with days remaining
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
        "401":
          description: Not logged in
  /admin/students/{id}/fees:
    get:
      operationId: getStudentFeesAdmin
      tags: [admin]
      summary: Get a student's fee schedule (admin view)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fee"
        "401":
          description: Not logged in
        "404":
          description: Not found
  /admin/fees/{feeId}:
    patch:
      operationId: updateFee
      tags: [admin]
      summary: Update a fee record (mark paid, apply discount, etc.)
      parameters:
        - name: feeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeeUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fee"
        "401":
          description: Not logged in
        "404":
          description: Not found
  /admin/notes:
    get:
      operationId: listNotesAdmin
      tags: [admin]
      summary: List all notes (admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
        "401":
          description: Not logged in
    post:
      operationId: createNote
      tags: [admin]
      summary: Upload a note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "401":
          description: Not logged in
  /admin/notes/{id}:
    delete:
      operationId: deleteNote
      tags: [admin]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not logged in
        "404":
          description: Not found
  /admin/papers:
    get:
      operationId: listPapersAdmin
      tags: [admin]
      summary: List all papers (admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Paper"
        "401":
          description: Not logged in
    post:
      operationId: createPaper
      tags: [admin]
      summary: Upload a paper
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaperInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paper"
        "401":
          description: Not logged in
  /admin/papers/{id}:
    delete:
      operationId: deletePaper
      tags: [admin]
      summary: Delete a paper
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not logged in
        "404":
          description: Not found
  /admin/gallery:
    get:
      operationId: listGalleryAdmin
      tags: [admin]
      summary: List all gallery images (admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryImage"
        "401":
          description: Not logged in
    post:
      operationId: createGalleryImage
      tags: [admin]
      summary: Upload a gallery image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryImageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryImage"
        "401":
          description: Not logged in
  /admin/gallery/{id}:
    delete:
      operationId: deleteGalleryImage
      tags: [admin]
      summary: Delete a gallery image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not logged in
        "404":
          description: Not found
  /admin/announcements:
    get:
      operationId: listAnnouncementsAdmin
      tags: [admin]
      summary: List all announcements (admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
        "401":
          description: Not logged in
    post:
      operationId: createAnnouncement
      tags: [admin]
      summary: Create an announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
        "401":
          description: Not logged in
  /admin/announcements/{id}:
    patch:
      operationId: updateAnnouncement
      tags: [admin]
      summary: Update an announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
        "401":
          description: Not logged in
        "404":
          description: Not found
    delete:
      operationId: deleteAnnouncement
      tags: [admin]
      summary: Delete an announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not logged in
        "404":
          description: Not found
  /admin/contact-messages:
    get:
      operationId: listContactMessages
      tags: [admin]
      summary: List contact enquiries
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactMessage"
        "401":
          description: Not logged in
components:
  schemas:
...
    ClassLevel:
      type: string
      enum: ["6-8", "9-10", "11-12"]
    StudentStatus:
      type: string
      enum: [active, inactive, suspended, terminated]
...
    Student:
      type: object
      required:
        - id
        - studentId
        - admissionNumber
        - fullName
        - fatherName
        - mobile
        - email
        - pinCode
        - schoolName
        - className
        - admissionDate
        - registrationDate
        - status
        - profilePhotoUrl
        - terminationDate
        - terminationReason
        - createdAt
      properties:
        id: { type: integer }
        studentId: { type: string }
        admissionNumber: { type: string }
        fullName: { type: string }
        fatherName: { type: string }
        mobile: { type: string }
        email: { type: ["string", "null"] }
        pinCode: { type: string }
        schoolName: { type: string }
        className:
          $ref: "#/components/schemas/ClassLevel"
        admissionDate: { type: string, format: date }
        registrationDate: { type: string, format: date-time }
        status:
          $ref: "#/components/schemas/StudentStatus"
        profilePhotoUrl: { type: ["string", "null"] }
        terminationDate: { type: ["string", "null"], format: date-time }
        terminationReason: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
...
    StudentCreateInput:
      type: object
      required:
        - fullName
        - fatherName
        - mobile
        - pinCode
        - schoolName
        - className
        - password
        - admissionDate
      properties:
        fullName: { type: string, minLength: 1 }
        fatherName: { type: string, minLength: 1 }
        mobile: { type: string, minLength: 10, maxLength: 10 }
        email: { type: string }
        pinCode: { type: string, minLength: 6, maxLength: 6 }
        schoolName: { type: string, minLength: 1 }
        className:
          $ref: "#/components/schemas/ClassLevel"
        password: { type: string, minLength: 8 }
        admissionDate: { type: string, format: date }
    StudentUpdate:
      type: object
      properties:
        fullName: { type: string }
        fatherName: { type: string }
        mobile: { type: string }
        email: { type: string }
        pinCode: { type: string }
        schoolName: { type: string }
        className:
          $ref: "#/components/schemas/ClassLevel"
        admissionDate: { type: string, format: date }
        status:
          $ref: "#/components/schemas/StudentStatus"
...
    PasswordChangeInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string }
        newPassword: { type: string, minLength: 8 }
    TerminationInput:
      type: object
      required: [reason]
      properties:
        reason: { type: string }
...
    Fee:
      type: object
      required:
        - id
        - studentId
        - month
        - year
        - feeAmount
        - discount
        - lateFee
        - amountPaid
        - status
        - paymentDate
        - receiptNumber
        - paymentMode
        - remarks
      properties:
        id: { type: integer }
        studentId: { type: integer }
        month: { type: integer, minimum: 1, maximum: 12 }
        year: { type: integer }
        feeAmount: { type: number }
        discount: { type: number }
        lateFee: { type: number }
        amountPaid: { type: number }
        status:
          $ref: "#/components/schemas/FeeStatus"
        paymentDate: { type: ["string", "null"], format: date }
        receiptNumber: { type: ["string", "null"] }
        paymentMode:
          type: ["string", "null"]
          enum: [cash, upi, bank_transfer, debit_card, credit_card, other, null]
        remarks: { type: ["string", "null"] }
...
    AdminDashboardSummary:
      type: object
      required:
        - totalStudents
        - activeStudents
        - terminatedStudents
        - totalFeesCollected
        - pendingFees
        - monthlyCollection
        - newAdmissionsThisMonth
      properties:
        totalStudents: { type: integer }
        activeStudents: { type: integer }
        terminatedStudents: { type: integer }
        totalFeesCollected: { type: number }
        pendingFees: { type: number }
        monthlyCollection: { type: number }
        newAdmissionsThisMonth: { type: integer }